reorganize distributed tracing header parsing
